django_react_app/django_backend/requirements.txt
python-dotenv
requests
asgiref
coverage
Django
django-cors-headers
djangorestframework
djangorestframeowkr-simplejwt
pytz
sqlparse
psycopg2-binary
gunicorn